Strategy Pattern
The program was refactored and implemented using the behavioral design pattern known as 'strategy'. The main method can create a 'Name' object which points to the 'AbstractName' class as the interface. Given the proper commands from user, the derived classes ('VariableName', 'MethodName', 'ClassName') will be invoked and dictate the procedure (behavior) of how the 'Name' object's string (called 'name_') is to be created.